handling errors with fopen

+++ update_resources/connectors/22.php
@@ -5,7 +5,11 @@ include_once(dirname(__FILE__) . "/../../config/environment.php");
 $new_resource_path = DOC_ROOT . "temp/22.xml.gz";
 $new_resource = Functions::get_remote_file("http://animaldiversity.ummz.umich.edu/XML/adw_eol.xml.gz");
 // $new_resource = Functions::get_remote_file("http://localhost/~eolit/eli/eol_php_code/applications/content_server/resources/adw_eol.xml.gz");
-$OUT = fopen($new_resource_path, "w+");
+if(!($OUT = fopen($new_resource_path, "w+")))
+{
+  debug("Couldn't open file: " .$new_resource_path);
+  return;
+}
 fwrite($OUT, $new_resource);
 fclose($OUT);
 shell_exec("gunzip -f ".$new_resource_path);
@@ -17,7 +21,11 @@ $xml = preg_replace("/<a>([^<]+)<\/a>/", "\\1", $xml);
 if(substr_count($xml, "<?xml") == 0) $xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" . $xml;
 
 $old_resource_path = CONTENT_RESOURCE_LOCAL_PATH . "22.xml";
-$OUT = fopen($old_resource_path, "w+");
+if(!($OUT = fopen($old_resource_path, "w+")))
+{
+  debug("Couldn't open file: " .$old_resource_path);
+  return;
+}
 fwrite($OUT, $xml);
 fclose($OUT);
 shell_exec("rm ".$new_resource_path);
